+++ src/Manager/CacheManager.php
@@ -9,14 +9,23 @@ class CacheManager {
     protected $defaultDriver = 'memory';
     /** @var array **/
     protected $pools;
-    
+    /** @var array **/
+    protected $options;
+
+    /**
+     * @param array $options
+     */
+    public function __construct($options = []) {
+        $this->options = $options;
+    }
+
     /**
      * @param \Psr\Cache\CacheItemInterface $item
      */
     public function save(CacheItemInterface $item) {
         $this->getItemPool($item->getDriver())->save($item);
     }
-    
+
     /**
      * @param string $key
      * @param string $driver
@@ -25,26 +34,26 @@ class CacheManager {
      */
     public function getItem($key, $driver = null, $ttl = null) {
         $finalDriver = ($driver !== null) ? $driver : $this->defaultDriver;
-        
+
         $pool = $this->getItemPool($finalDriver);
         $item = $pool->getItem($key);
-        
+
         // In this case, the pool returned a new CacheItem
         if($item->get() === null) {
             $item->expiresAfter($ttl);
         }
         return $item;
     }
-    
+
     /**
-     * 
+     *
      * @param string $driver
      * @return Psr\Cache\CacheItemPoolInterface
      */
     public function getItemPool($driver) {
         if(!isset($this->pools[$driver])) {
             $poolClass = 'PhpAbac\\Cache\\Pool\\' . ucfirst($driver) . 'CacheItemPool';
-            $this->pools[$driver] = new $poolClass();
+            $this->pools[$driver] = new $poolClass($this->options);
         }
         return $this->pools[$driver];
     }